Fixed issues pointed out by static analysis.

+++ src/libxsmm_math.c
@@ -120,7 +120,7 @@ LIBXSMM_API int libxsmm_matdiff(libxsmm_datatype datatype, libxsmm_blasint m, li
       const char *const env = getenv("LIBXSMM_DUMP");
       if (0 != env && 0 != *env && (('0' < *env && '9' >= *env) || '0' != *env)) {
         const char *const defaultname = ('0' < *env && '9' >= *env) ? "libxsmm_dump" : env;
-        const libxsmm_mhd_elemtype type_src = (libxsmm_mhd_elemtype)((int)datatype);
+        const libxsmm_mhd_elemtype type_src = (libxsmm_mhd_elemtype)datatype;
         const libxsmm_mhd_elemtype type_dst = LIBXSMM_MAX(LIBXSMM_MHD_ELEMTYPE_U8, type_src);
         char filename[256];
         size_t size[2], pr[2]; size[0] = (size_t)mm; size[1] = (size_t)nn; pr[0] = (size_t)ldr; pr[1] = (size_t)nn;
